Renamed domHelpers -> overlayHelpers
Name now better reflects responsibility
